[util] Fold ansi-print.cc into ansi-print.hh
